requirements.txt
appdirs==1.4.3
click==6.7
croniter==0.3.16
Django==1.10.6
django-extensions==1.7.7
django-rq==0.9.5
future==0.16.0
Jinja2==2.9.5
lxml==3.7.3
MarkupSafe==1.0
oauthlib==2.0.2
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
python-dateutil==2.6.0
python-twitter==3.2.1
redis==2.10.5
requests==2.13.0
requests-oauthlib==0.8.0
rq==0.7.1
rq-scheduler==0.7.0
six==1.10.0
pyenchant==1.6.8
tweepy==3.5.0
